% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ia_files.R
\name{ia_files}
\alias{ia_files}
\title{Access the list of files associated with an Internet Archive item}
\usage{
ia_files(items)
}
\arguments{
\item{items}{A list describing an Internet Archive items returned from
the API.}
}
\value{
A list containing the files as a list of character vectors.
}
\description{
Access the list of files associated with an Internet Archive item
}
\examples{
\dontrun{
ats_query <- c("publisher" = "american tract society")
ids       <- ia_search(ats_query, num_results = 3)
items     <- ia_get_items(ids)
files     <- ia_files(items)
files
}
}
