% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ia_get_item.R
\name{ia_get_item}
\alias{ia_get_item}
\title{Get the metadata for an Internet Archive item}
\usage{
ia_get_item(item_id)
}
\arguments{
\item{item_id}{A character vector containing the ID for an Internet Archive
item.}
}
\value{
A list containing the metadata returned by the API.
}
\description{
Get the metadata for an Internet Archive item
}

