% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ia_keyword_search.R
\name{ia_keyword_search}
\alias{ia_keyword_search}
\title{Perform an simple keyword search of the Internet Archive.}
\usage{
ia_keyword_search(keywords, num_results = 5, page = 1, print_total = TRUE)
}
\arguments{
\item{keywords}{The keywords to search for.}

\item{num_results}{The number of results to return per page.}

\item{page}{When results are paged, which page of results to return.}

\item{print_total}{Should the total number of results for this query be
printed as a message?}
}
\value{
A character vector of Internet Archive item IDs.
}
\description{
Perform an simple keyword search of the Internet Archive.
}
\examples{
ia_keyword_search("isaac hecker", num_results = 20)
}

