% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ia_list_fields.R
\name{ia_list_fields}
\alias{ia_list_fields}
\title{List accepted metadata fields}
\usage{
ia_list_fields()
}
\value{
A list of the accepted metadata fields
}
\description{
List accepted metadata fields
}
\examples{
ia_list_fields()
}
