% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ia_metadata.R
\name{ia_metadata}
\alias{ia_metadata}
\title{Access the item metadata from an Internet Archive item}
\usage{
ia_metadata(items)
}
\arguments{
\item{items}{A list object describing an Internet Archive items returned from
the API.}
}
\value{
A data frame containing the metadata, with columns \code{id} for the
  item identifier, \code{field} for the name of the metadata field, and
  \code{value} for the metadata values.
}
\description{
Access the item metadata from an Internet Archive item
}

