% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ia_wayback.R
\name{ia_wayback_search}
\alias{ia_wayback_search}
\title{Search the wayback machine}
\usage{
ia_wayback_search(wayback_url, timestamp = NULL)
}
\arguments{
\item{wayback_url}{The url for which to get a snapshot.}

\item{timestamp}{timestamp in format YYYYMMDDhhmmss. YYYYMMDD will also work.}
}
\value{
List containing the closest snapshot.
}
\description{
Provide and url and timestamp and get a url back to the closest snapshot the wayback machine crawled.
}
\examples{
#  wayback <- ia_wayback_search("http://lincolnmullen.com/#software",'20160701005604')
}
\references{
See the documentation on the Internet Archive's
\href{https://archive.org/help/wayback_api.php}{wayback page}.
}

