% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/internetarchive-package.R
\name{internetarchive}
\alias{internetarchive}
\title{Client for the Internet Archive API}
\description{
This client permits you to search (\link{ia_search}), retrieve item metadata
(\link{ia_metadata}) and associated files (\link{ia_files}), and download
files (\link{ia_files}) in a pipeable interface.
}

