% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ia_item_id.R
\name{ia_item_id}
\alias{ia_item_id}
\title{Access the item IDs from an Internet Archive items}
\usage{
ia_item_id(item)
}
\arguments{
\item{item}{A list describing an Internet Archive items returned from
the API. This argument is vectorized.}
}
\value{
A character vector containing the item IDs.
}
\description{
Access the item IDs from an Internet Archive items
}
\examples{
ats_query <- c("publisher" = "american tract society")
ids       <- ia_search(ats_query, num_results = 3)
items     <- ia_get_items(ids)
ia_item_id(items)
}

